+++ lib/unisec/cli/blocks.rb
@@ -52,7 +52,7 @@ module Unisec
           option :with_count, default: 'false', values: %w[true false],
                               desc: "calculate block's range size & char count?"
 
-          # List code points matching a Unicode property
+          # Display a block matching a decimal code point, standardized hexadecimal codepoint, string character or block name
           # @param block_arg [Integer|String] Decimal code point or standardized hexadecimal codepoint or string character (only one, so be careful with emojis, composed or joint characters using several units) or directly look for the block name (case insensitive).
           def call(block_arg: nil, **options)
             block_arg = block_arg.to_i if /\A\d+\Z/.match?(block_arg) # cast decimal string to integer